+++ src/transaction/Beef.ts
@@ -4,6 +4,7 @@ import ChainTracker from './ChainTracker.js'
 import BeefTx from './BeefTx.js'
 import { Reader, Writer, toHex, toArray, verifyNotNull } from '../primitives/utils.js'
 import { hash256 } from '../primitives/Hash.js'
+import { ReaderUint8Array } from '../primitives/ReaderUint8Array.js'
 
 const BufferCtor =
   typeof globalThis !== 'undefined' ? (globalThis as any).Buffer : undefined
@@ -117,6 +118,30 @@ export class Beef {
       this.rawBytesCache = writer.toUint8Array()
     }
     return this.rawBytesCache
+  }
+
+  private getBeefForAtomic (txid: string): { beef: Beef, writer: Writer } {
+    if (this.needsSort) {
+      this.sortTxs()
+    }
+    const tx = this.findTxid(txid)
+    if (tx == null) {
+      throw new Error(`${txid} does not exist in this Beef`)
+    }
+
+    // If the transaction is not the last one, clone and modify
+    const beef = (this.txs[this.txs.length - 1] === tx) ? this : this.clone()
+
+    if (beef !== this) {
+      const i = this.txs.findIndex((t) => t.txid === txid)
+      beef.txs.splice(i + 1)
+    }
+
+    const writer = new Writer()
+    writer.writeUInt32LE(ATOMIC_BEEF)
+    writer.writeReverse(toArray(txid, 'hex'))
+
+    return { beef, writer }
   }
 
   /**
@@ -308,7 +333,7 @@ export class Beef {
    * @param bumpIndex Optional. If a number, must be valid index into bumps array.
    * @returns txid of rawTx
    */
-  mergeRawTx (rawTx: number[], bumpIndex?: number): BeefTx {
+  mergeRawTx (rawTx: number[] | Uint8Array, bumpIndex?: number): BeefTx {
     this.markMutated(true)
     const newTx: BeefTx = new BeefTx(rawTx, bumpIndex)
     this.removeExistingTxid(newTx.txid)
@@ -394,8 +419,8 @@ export class Beef {
     return beefTx
   }
 
-  mergeBeef (beef: number[] | Beef): void {
-    const b: Beef = Array.isArray(beef) ? Beef.fromBinary(beef) : beef
+  mergeBeef (beef: Beef | number[] | Uint8Array): void {
+    const b: Beef = (beef instanceof Beef) ? beef : Beef.fromBinary(beef)
 
     for (const bump of b.bumps) {
       this.mergeBump(bump)
@@ -566,11 +591,16 @@ export class Beef {
   /**
    * Returns a binary array representing the serialized BEEF
    * @returns A binary array representing the BEEF
+   * @returns An array of byte values containing binary serialization of the BEEF
    */
   toBinary (): number[] {
     return Array.from(this.getSerializedBytes())
   }
 
+  /**
+   * Returns a binary array representing the serialized BEEF
+   * @returns A Uint8Array containing binary serialization of the BEEF
+   */
   toUint8Array (): Uint8Array {
     return this.getSerializedBytes()
   }
@@ -586,28 +616,29 @@ export class Beef {
    * @returns serialized contents of this Beef with AtomicBEEF prefix.
    */
   toBinaryAtomic (txid: string): number[] {
-    if (this.needsSort) {
-      this.sortTxs()
-    }
-    const tx = this.findTxid(txid)
-    if (tx == null) {
-      throw new Error(`${txid} does not exist in this Beef`)
-    }
-
-    // If the transaction is not the last one, clone and modify
-    const beef = (this.txs[this.txs.length - 1] === tx) ? this : this.clone()
-
-    if (beef !== this) {
-      const i = this.txs.findIndex((t) => t.txid === txid)
-      beef.txs.splice(i + 1)
-    }
-
-    const writer = new Writer()
-    writer.writeUInt32LE(ATOMIC_BEEF)
-    writer.writeReverse(toArray(txid, 'hex'))
+    const { beef, writer } = this.getBeefForAtomic(txid)
     beef.toWriter(writer)
-
     return writer.toArray()
+  }
+
+  /**
+   * Serialize this Beef as AtomicBEEF.
+   *
+   * `txid` must exist
+   *
+   * after sorting, if txid is not last txid, creates a clone and removes newer txs
+   *
+   * @param txid
+   * @returns serialized contents of this Beef with AtomicBEEF prefix.
+   */
+  toUint8ArrayAtomic (txid: string): Uint8Array {
+    const { beef, writer } = this.getBeefForAtomic(txid)
+    const beefUint8 = beef.getSerializedBytes()
+    const prefix = writer.toUint8Array()
+    const atomic = new Uint8Array(prefix.length + beefUint8.length)
+    atomic.set(prefix, 0)
+    atomic.set(beefUint8, prefix.length)
+    return atomic
   }
 
   /**
@@ -627,7 +658,7 @@ export class Beef {
     return hex
   }
 
-  static fromReader (br: Reader): Beef {
+  static fromReader (br: Reader | ReaderUint8Array): Beef {
     let version = br.readUInt32LE()
     let atomicTxid: string | undefined
     if (version === ATOMIC_BEEF) {
@@ -657,11 +688,11 @@ export class Beef {
 
   /**
    * Constructs an instance of the Beef class based on the provided binary array
-   * @param bin The binary array from which to construct BEEF
+   * @param bin The binary array or Uint8Array from which to construct BEEF
    * @returns An instance of the Beef class constructed from the binary data
    */
-  static fromBinary (bin: number[]): Beef {
-    const br = new Reader(bin)
+  static fromBinary (bin: number[] | Uint8Array): Beef {
+    const br = ReaderUint8Array.makeReader(bin)
     return Beef.fromReader(br)
   }