+++ src/transaction/Transaction.ts
@@ -3,7 +3,7 @@ import TransactionInput from './TransactionInput.js'
 import TransactionOutput from './TransactionOutput.js'
 import UnlockingScript from '../script/UnlockingScript.js'
 import LockingScript from '../script/LockingScript.js'
-import { Reader, Writer, toHex, toArray } from '../primitives/utils.js'
+import { Reader, Writer, toHex, toArray, ReaderUint8Array, toUint8Array } from '../primitives/utils.js'
 import { hash256 } from '../primitives/Hash.js'
 import FeeModel from './FeeModel.js'
 import LivePolicy from './fee-models/LivePolicy.js'
@@ -110,7 +110,7 @@ export default class Transaction {
    * @param txid Optional TXID of the transaction to retrieve from the BEEF data.
    * @returns An anchored transaction, linked to its associated inputs populated with merkle paths.
    */
-  static fromBEEF (beef: number[], txid?: string): Transaction {
+  static fromBEEF (beef: BEEF, txid?: string): Transaction {
     const { tx } = Transaction.fromAnyBeef(beef, txid)
     return tx
   }
@@ -122,7 +122,7 @@ export default class Transaction {
    * @param beef A binary representation of an Atomic BEEF structure.
    * @returns The subject transaction, linked to its associated inputs populated with merkle paths.
    */
-  static fromAtomicBEEF (beef: number[]): Transaction {
+  static fromAtomicBEEF (beef: BEEF): Transaction {
     const { tx, txid, beef: b } = Transaction.fromAnyBeef(beef)
     if (txid !== b.atomicTxid) {
       if (b.atomicTxid != null) {
@@ -134,7 +134,7 @@ export default class Transaction {
     return tx
   }
 
-  private static fromAnyBeef (beef: number[], txid?: string): { tx: Transaction, beef: Beef, txid: string } {
+  private static fromAnyBeef (beef: BEEF, txid?: string): { tx: Transaction, beef: Beef, txid: string } {
     const b = Beef.fromBinary(beef)
     if (b.txs.length < 1) {
       throw new Error('beef must include at least one transaction.')
@@ -156,8 +156,8 @@ export default class Transaction {
    * @param ef A binary representation of a transaction in EF format.
    * @returns An extended transaction, linked to its associated inputs by locking script and satoshis amounts only.
    */
-  static fromEF (ef: number[]): Transaction {
-    const br = new Reader(ef)
+  static fromEF (ef: number[] | Uint8Array): Transaction {
+    const br = ReaderUint8Array.makeReader(ef)
     const version = br.readUInt32LE()
     if (toHex(br.read(6)) !== '0000000000ef') { throw new Error('Invalid EF marker') }
     const inputsLength = br.readVarIntNum()
@@ -244,7 +244,7 @@ export default class Transaction {
     return { inputs, outputs }
   }
 
-  static fromReader (br: Reader): Transaction {
+  static fromReader (br: Reader | ReaderUint8Array): Transaction {
     const version = br.readUInt32LE()
     const inputsLength = br.readVarIntNum()
     const inputs: TransactionInput[] = []
@@ -285,10 +285,10 @@ export default class Transaction {
    * @param {number[]} bin - The binary array representation of the transaction.
    * @returns {Transaction} - A new Transaction instance.
    */
-  static fromBinary (bin: number[]): Transaction {
+  static fromBinary (bin: number[] | Uint8Array): Transaction {
     const copy = bin.slice()
     const rawBytes = Uint8Array.from(copy)
-    const br = new Reader(copy)
+    const br = new ReaderUint8Array(rawBytes)
     const tx = Transaction.fromReader(br)
     tx.rawBytesCache = rawBytes
     return tx
@@ -302,15 +302,11 @@ export default class Transaction {
    * @returns {Transaction} - A new Transaction instance.
    */
   static fromHex (hex: string): Transaction {
-    const bin = toArray(hex, 'hex')
-    const rawBytes = Uint8Array.from(bin)
-    const br = new Reader(bin)
+    const rawBytes = toUint8Array(hex, 'hex')
+    const br = new ReaderUint8Array(rawBytes)
     const tx = Transaction.fromReader(br)
     tx.rawBytesCache = rawBytes
-    tx.hexCache =
-      BufferCtor != null
-        ? BufferCtor.from(rawBytes).toString('hex')
-        : toHex(bin)
+    tx.hexCache = toHex(rawBytes)
     return tx
   }
 
@@ -322,7 +318,7 @@ export default class Transaction {
    * @returns {Transaction} - A new Transaction instance.
    */
   static fromHexEF (hex: string): Transaction {
-    return Transaction.fromEF(toArray(hex, 'hex'))
+    return Transaction.fromEF(toUint8Array(hex, 'hex'))
   }
 
   /**